+++ cider-inspector.el
@@ -81,7 +81,8 @@ The page size can be also changed interactively within the inspector."
   (set-syntax-table clojure-mode-syntax-table)
   (setq buffer-read-only t)
   (setq-local electric-indent-chars nil)
-  (setq-local truncate-lines t))
+  (when cider-special-mode-truncate-lines
+    (setq-local truncate-lines t)))
 
 ;;;###autoload
 (defun cider-inspect-last-sexp ()